+++ project/src/components/const.ts
@@ -1,0 +1,12 @@
+export enum AppRoute {
+    MainPage = '/',
+    Login = '/login',
+    Favorites = '/favorites',
+    Room = '/offer/:id',
+}
+
+export enum AuthorizationStatus {
+    Auth = 'AUTH',
+    NoAuth = 'NO_AUTH',
+    Unknown = 'UNKNOWN',
+}
